Beginnings of stuff with client ID
Main part is the getClientID which the client uses to know it's ID; It
first looks if it has an idfile clientid.txt and if it does, that is
what it will use to communicate with the server.  However, if the file
does not exist, it will use the createClientID method of the server
interface to get an ID from the server.

For now, the server returns -1 as the client ID but it will maintain
a list of IDs.

Also, I added an ID to the response class because why not.  It might be
useful.