Adding client IDs to a file in the server
When the server starts, it reads the IDs into an array list

When a demand to create a client id is made, it looks in that list and
finds the maximum element.  It then uses max+1 as the new ID.

The ID is added to the arraylist and appended to the file at the same
time.  This is for robustness in case the server crashes.

Also, the server shows the list of IDs for demonstration purposes.